requirements.txt
numpy==1.26.4
pandas==2.2.2
polars==0.20.26
scikit-learn==1.4.2
matplotlib==3.8.4
scipy==1.13.0
seaborn==0.13.2
hiplot==0.1.33
ipython==8.24.0
statannotations==0.5.0
kaggle
antropy
tqdm
umap-learn==0.5.8
altair